refactor(packaging): view/fragment
