+++ app/src/main/java/com/example/ezvault/view/fragment/AddItemFragment.java
@@ -1,4 +1,4 @@
-package com.example.ezvault;
+package com.example.ezvault.view.fragment;
 
 
 
@@ -36,6 +36,8 @@ import androidx.navigation.Navigation;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.ezvault.PhotoAdapter;
+import com.example.ezvault.R;
 import com.example.ezvault.camera.GalleryAction;
 import com.example.ezvault.database.FirebaseBundle;
 import com.example.ezvault.database.ImageDAO;
@@ -50,6 +52,7 @@ import com.example.ezvault.model.SerialPredictor;
 import com.example.ezvault.model.Tag;
 import com.example.ezvault.model.User;
 import com.example.ezvault.textwatchers.NonEmptyTextWatcher;
+import com.example.ezvault.upcAPI;
 import com.example.ezvault.utils.FileUtils;
 import com.example.ezvault.utils.TaskUtils;
 import com.example.ezvault.utils.UserManager;
@@ -83,7 +86,7 @@ import java.util.stream.Collectors;
 @AndroidEntryPoint
 public class AddItemFragment extends Fragment {
 
-    private final upcAPI upcLookup = new upcAPI();
+    private final com.example.ezvault.upcAPI upcLookup = new upcAPI();
     private Button createButton;
 
     private AutoCompleteTextView itemSerial;
