+++ app/src/main/java/com/example/ezvault/view/fragment/LoginFragment.java
@@ -1,4 +1,4 @@
-package com.example.ezvault;
+package com.example.ezvault.view.fragment;
 
 import android.os.Bundle;
 import android.util.Log;
@@ -15,6 +15,7 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
 
+import com.example.ezvault.R;
 import com.example.ezvault.authentication.authentication.AuthenticationHandler;
 import com.example.ezvault.authentication.authentication.EmailPasswordAuthenticationStrategy;
 import com.example.ezvault.database.FirebaseBundle;
