+++ app/src/main/java/com/example/ezvault/view/fragment/TagsFragment.java
@@ -1,4 +1,4 @@
-package com.example.ezvault.view;
+package com.example.ezvault.view.fragment;
 
 import android.os.Bundle;
