+++ app/src/main/res/layout/fragment_filter.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.FilterFragment">
+    tools:context=".view.fragment.FilterFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
