+++ app/src/main/res/layout/fragment_view_item_details.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".EditItemDetails">
+    tools:context=".view.fragment.EditItemDetails">
 
     <ScrollView
         android:layout_width="match_parent"
